requirements.txt
asgiref==3.8.1
asttokens==2.4.1
attrs==23.2.0
black==24.4.0
click==8.1.7
colorama==0.4.6
decorator==5.1.1
dj-database-url==2.1.0
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
executing==2.0.1
inflection==0.5.1
ipdb==0.13.13
ipython==8.23.0
jedi==0.19.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
matplotlib-inline==0.1.7
mypy-extensions==1.0.0
packaging==24.0
parso==0.8.4
pathspec==0.12.1
platformdirs==4.2.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pure-eval==0.2.2
Pygments==2.17.2
PyJWT==2.8.0
python-dotenv==1.0.1
PyYAML==6.0.1
referencing==0.35.0
rpds-py==0.18.0
six==1.16.0
sqlparse==0.5.0
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
wcwidth==0.2.13
